+++ libraries/AP_ESC_Telem/AP_ESC_Telem_Backend.cpp
@@ -31,7 +31,7 @@ AP_ESC_Telem_Backend::AP_ESC_Telem_Backend() {
 }
 
 // callback to update the rpm in the frontend, should be called by the driver when new data is available
-void AP_ESC_Telem_Backend::update_rpm(const uint8_t esc_index, const uint16_t new_rpm, const float error_rate) {
+void AP_ESC_Telem_Backend::update_rpm(const uint8_t esc_index, const float new_rpm, const float error_rate) {
     _frontend->update_rpm(esc_index, new_rpm, error_rate);
 }
